requirements.txt
requests-html
lxml_html_clean
python-dotenv

pandas